Replace PathMap.normalisePath with the one from alhadis.utils
`PathMap` currently has a different `normalisePath` implementation from
the one that the `Resource` class uses, leading to a mismatch between
`PathMap` keys and resource paths. This can lead to bugs where a
resource is marked as disposed but not removed from the `PathMap`;
attempting to fetch it again leads to exceptions from the null `icon`.